requirements.txt
beautifulsoup4>=4.13.3
httpx>=0.28.1
pydantic-settings>=2.8.1
pydantic>=2.10.6
elasticsearch>=8.11.0
asyncio>=3.4.3
aiofiles>=23.2.1
openai>=1.12.0
lxml>=5.3.1
pandas
python-dotenv